fix: autoconnect and allow disconnect
